# apps/client/latest_openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
...
    "/admin/permissions": {
      "get": {
        "tags": ["admin"],
        "summary": "List Permissions",
        "operationId": "list_permissions_admin_permissions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoutePermissionResponse"
                  },
                  "type": "array",
                  "title": "Response List Permissions Admin Permissions Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["admin"],
        "summary": "Create Permission",
        "operationId": "create_permission_admin_permissions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RoutePermissionCreate" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePermissionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/admin/permissions/{permission_id}": {
      "put": {
        "tags": ["admin"],
        "summary": "Update Permission",
        "operationId": "update_permission_admin_permissions__permission_id__put",
        "parameters": [
          {
            "name": "permission_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Permission Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RoutePermissionUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePermissionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["admin"],
        "summary": "Delete Permission",
        "operationId": "delete_permission_admin_permissions__permission_id__delete",
        "parameters": [
          {
            "name": "permission_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Permission Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/admin/role-rules": {
      "get": {
        "tags": ["admin"],
        "summary": "List Role Rules",
        "operationId": "list_role_rules_admin_role_rules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleAssignmentRuleResponse"
                  },
                  "type": "array",
                  "title": "Response List Role Rules Admin Role Rules Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["admin"],
        "summary": "Create Role Rule",
        "operationId": "create_role_rule_admin_role_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleAssignmentRuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleAssignmentRuleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/admin/role-rules/{rule_id}": {
      "put": {
        "tags": ["admin"],
        "summary": "Update Role Rule",
        "operationId": "update_role_rule_admin_role_rules__rule_id__put",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Rule Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleAssignmentRuleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleAssignmentRuleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["admin"],
        "summary": "Delete Role Rule",
        "operationId": "delete_role_rule_admin_role_rules__rule_id__delete",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Rule Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/admin/users/{user_id}": {
      "get": {
        "tags": ["admin"],
        "summary": "Get User",
        "operationId": "get_user_admin_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "User Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserAdminResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["admin"],
        "summary": "Update User",
        "operationId": "update_user_admin_users__user_id__put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "User Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserAdminResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["admin"],
        "summary": "Delete User",
        "operationId": "delete_user_admin_users__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "User Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/admin/terms": {
      "get": {
        "tags": ["admin"],
        "summary": "List Terms",
        "operationId": "list_terms_admin_terms_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": { "$ref": "#/components/schemas/TermResponse" },
                  "type": "array",
                  "title": "Response List Terms Admin Terms Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["admin"],
        "summary": "Create Term",
        "operationId": "create_term_admin_terms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TermCreate" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TermResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/admin/terms/{term_id}": {
      "put": {
        "tags": ["admin"],
        "summary": "Update Term",
        "operationId": "update_term_admin_terms__term_id__put",
        "parameters": [
          {
            "name": "term_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Term Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TermUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TermResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["admin"],
        "summary": "Delete Term",
        "operationId": "delete_term_admin_terms__term_id__delete",
        "parameters": [
          {
            "name": "term_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Term Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsentResponse": {
        "properties": {
          "term_id": { "type": "integer", "title": "Term Id" },
          "agreed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Agreed At"
          }
        },
        "type": "object",
        "required": ["term_id", "agreed_at"],
        "title": "ConsentResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoleAssignmentRuleCreate": {
        "properties": {
          "rule_type": { "$ref": "#/components/schemas/RuleType" },
          "rule_value": {
            "additionalProperties": true,
            "type": "object",
            "title": "Rule Value"
          },
          "assigned_role": { "type": "string", "title": "Assigned Role" },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "default": 100
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": ["rule_type", "rule_value", "assigned_role"],
        "title": "RoleAssignmentRuleCreate"
      },
      "RoleAssignmentRuleResponse": {
        "properties": {
          "rule_type": { "$ref": "#/components/schemas/RuleType" },
          "rule_value": {
            "additionalProperties": true,
            "type": "object",
            "title": "Rule Value"
          },
          "assigned_role": { "type": "string", "title": "Assigned Role" },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "default": 100
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "id": { "type": "integer", "title": "Id" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "rule_type",
          "rule_value",
          "assigned_role",
          "id",
          "created_at"
        ],
        "title": "RoleAssignmentRuleResponse"
      },
      "RoleAssignmentRuleUpdate": {
        "properties": {
          "rule_value": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Rule Value"
          },
          "assigned_role": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Assigned Role"
          },
          "priority": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Priority"
          },
          "is_active": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "RoleAssignmentRuleUpdate"
      },
      "RoutePermissionCreate": {
        "properties": {
          "path": { "type": "string", "title": "Path" },
          "method": { "type": "string", "title": "Method" },
          "is_public": {
            "type": "boolean",
            "title": "Is Public",
            "default": false
          },
          "roles": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Roles",
            "default": []
          }
        },
        "type": "object",
        "required": ["path", "method"],
        "title": "RoutePermissionCreate"
      },
      "RoutePermissionResponse": {
        "properties": {
          "path": { "type": "string", "title": "Path" },
          "method": { "type": "string", "title": "Method" },
          "is_public": {
            "type": "boolean",
            "title": "Is Public",
            "default": false
          },
          "roles": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Roles",
            "default": []
          },
          "id": { "type": "integer", "title": "Id" }
        },
        "type": "object",
        "required": ["path", "method", "id"],
        "title": "RoutePermissionResponse"
      },
      "RoutePermissionUpdate": {
        "properties": {
          "is_public": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Public"
          },
          "roles": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Roles"
          }
        },
        "type": "object",
        "title": "RoutePermissionUpdate"
      },
      "RuleType": {
        "type": "string",
        "enum": ["email_pattern", "email_list"],
        "title": "RuleType"
      },
...
      "TermCreate": {
        "properties": {
          "type": { "type": "string", "title": "Type" },
          "version": { "type": "string", "title": "Version" },
          "content": { "type": "string", "title": "Content" },
          "is_required": {
            "type": "boolean",
            "title": "Is Required",
            "default": false
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": false
          }
        },
        "type": "object",
        "required": ["type", "version", "content"],
        "title": "TermCreate"
      },
      "TermResponse": {
        "properties": {
          "type": { "type": "string", "title": "Type" },
          "version": { "type": "string", "title": "Version" },
          "content": { "type": "string", "title": "Content" },
          "is_required": {
            "type": "boolean",
            "title": "Is Required",
            "default": false
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": false
          },
          "id": { "type": "integer", "title": "Id" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": ["type", "version", "content", "id", "created_at"],
        "title": "TermResponse"
      },
      "TermUpdate": {
        "properties": {
          "content": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Content"
          },
          "is_required": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Required"
          },
          "is_active": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "TermUpdate"
      },
      "UserAdminResponse": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "google_sub": { "type": "string", "title": "Google Sub" },
          "email": { "type": "string", "title": "Email" },
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "picture": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Picture"
          },
          "roles": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Roles"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "consents": {
            "items": { "$ref": "#/components/schemas/ConsentResponse" },
            "type": "array",
            "title": "Consents",
            "default": []
          }
        },
        "type": "object",
        "required": ["id", "google_sub", "email", "roles", "created_at"],
        "title": "UserAdminResponse"
      },
...
      "UserUpdate": {
        "properties": {
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "picture": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Picture"
          },
          "roles": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Roles"
          }
        },
        "type": "object",
        "title": "UserUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}